CRR Binomial Tree Option Pricer: From European to American Exercise
This Python file implements the Cox-Ross-Rubinstein (CRR) binomial tree model for pricing options. It is designed as an educational tool to demonstrate the fundamental building blocks of a lattice model and its extension from evaluating European-style options (which can only be exercised at expiration) to American-style options (which can be exercised at any time before expiration).